rename afterFlush to nextTick, schedule update if necessary
