+++ test/runtime/samples/lifecycle-next-tick/main.html
@@ -1,0 +1,24 @@
+<script>
+	import { nextTick } from 'svelte';
+
+	export let snapshots = [];
+
+	let count = 0;
+	let buttons = [];
+
+	function increment() {
+		count += 1;
+		log();
+	}
+
+	function log() {
+		snapshots.push(`before ${buttons[0].textContent}`);
+
+		nextTick(() => {
+			snapshots.push(`after ${buttons[0].textContent}`);
+		});
+	}
+</script>
+
+<button bind:this={buttons[0]} on:click={increment}>{count}</button>
+<button bind:this={buttons[1]} on:click={log}>{count}</button>
